move CpuIdDump to the crate
